sip-types: rewrite host parsing
The previous implementation was hardly spec compliant
and sometimes mistook hostnames for ip-addresses.

This commit implements a more nom-style parser which
completely tries to parse first ipv6, then ipv4 and
then the hostname. As it turns out the SIP ABNF for
hostnames is too strict and still requires the toplabel
to start with a letter. This implentation ignores that
for now and allows any alphanumeric characters and hyphens
inside labels. It only checks that the label doesn't start
or ends with a hyphen.